Fixed warnings

+++ installer/dev/install.cpp
@@ -386,7 +386,7 @@ namespace WindowsAppRuntimeInstaller
 
                 // Restart Push Notifications Long Running Platform when ForceDeployment option is applied.
                 if (WI_IsFlagSet(options, WindowsAppRuntimeInstaller::Options::ForceDeployment) &&
-                    CompareStringOrdinal(package.id.c_str(), package.id.size() - 3, WAR_SINGLETON_X86_ID, package.id.size() - 3, TRUE) == CSTR_EQUAL)
+                    CompareStringOrdinal(package.id.c_str(), static_cast<int>(package.id.size() - 3), WAR_SINGLETON_X86_ID, static_cast<int>(package.id.size() - 3), TRUE) == CSTR_EQUAL)
                 {
                     RestartPushNotificationsLRP();
                 }
